Діаметр і товщина корпусу тепер приймають дробові мм (Numeric замість int)
Тестовий імпорт Casio показав, що виробники зазначають розмір годинника
дробовим числом (38.5мм, 44.2мм) — це стандарт, не помилка в даних.
Ціле поле обрізало дробову частину.

Міняю case_diameter_mm/case_thickness_mm на numeric(4,1) (той самий
підхід, що вже в products.price — точне десяткове число, без похибок
float) і синхронно правлю все, що з ними працює: ORM-модель, Pydantic-
схеми, query-параметри діапазонного фільтра, парсер CSV-імпорту
(parse_optional_int замінено на вже наявний parse_optional_decimal),
і хардкод int() у catalog_service.get_numeric_ranges, який обрізав межі
діапазону фільтра на сторінці каталогу.

+++ backend/app/schemas/admin_product.py
@@ -48,8 +48,8 @@ class AdminProductDetail(BaseModel):
     brand_id: int
     mechanism_type_id: int
     gender: Gender
-    case_diameter_mm: int | None
-    case_thickness_mm: int | None
+    case_diameter_mm: float | None
+    case_thickness_mm: float | None
     warranty_months: int | None
     package_contents: str | None
     is_active: bool
@@ -68,8 +68,8 @@ class ProductCreateRequest(BaseModel):
     brand_id: int
     mechanism_type_id: int
     gender: Gender
-    case_diameter_mm: int | None = None
-    case_thickness_mm: int | None = None
+    case_diameter_mm: float | None = None
+    case_thickness_mm: float | None = None
     warranty_months: int | None = None
     package_contents: str | None = None
     is_active: bool = True
@@ -86,8 +86,8 @@ class ProductUpdateRequest(BaseModel):
     brand_id: int | None = None
     mechanism_type_id: int | None = None
     gender: Gender | None = None
-    case_diameter_mm: int | None = None
-    case_thickness_mm: int | None = None
+    case_diameter_mm: float | None = None
+    case_thickness_mm: float | None = None
     warranty_months: int | None = None
     package_contents: str | None = None
     is_active: bool | None = None
